Added Classification Tree SVM
Build a model that predicts if a credit card transaction is fraudulent or not. You can model the problem as a binary classification problem. A transaction belongs to the positive class (1) if it is a fraud, otherwise it belongs to the negative class (0). Model a classification task using the Scikit-Learn and Snap ML Python APIs. Train Support Vector Machine and Decision Tree models using Scikit-Learn and Snap ML.